# api/docs/swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma Traceo — localização de pessoas desaparecidas no Brasil.",
        "title": "Traceo API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/homeless/{id}/matches": {
            "get": {
                "description": "Retorna candidatos de match para um homeless",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Listar matches de um homeless",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Homeless ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.MatchResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{id}": {
            "patch": {
                "description": "Confirmar ou rejeitar match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Atualizar status do match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.UpdateMatchStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_l3co_traceo-api_pkg_httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/missing/{id}/matches": {
            "get": {
                "description": "Retorna matches encontrados para um missing",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Listar matches de um desaparecido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Missing ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.MatchResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_l3co_traceo-api_pkg_httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.MatchResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "gemini_analysis": {
                    "type": "string"
                },
                "homeless_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "missing_id": {
                    "type": "string"
                },
                "reviewed_at": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.UpdateMatchStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}